extract ld lowering helper
